refactor(vscode-extension): fix Kilo Code agent call
Removed the `injectPromptDiagnosticWithCallback` utility and directly call the `kilo-code.newTask` command passing the prompt. I had originally misunderstood the utilitity. From my current understanding, it is used as a workaround in order to cause the agent to run a task without having direct acess to a `newTask` command, but please correct me if I'm misunderstanding it things!

WIth the the `kilo-code.newTask` there shouldn't be a need to inject diagnostics since Kilo should be able to read them itself.

Test plan: I haven't gotta a chance to test this with the stagewise extension quite yet because I don't have a good test env setup yet, but I can work on that!